+++ changelog.md
@@ -1,3 +1,14 @@
+# 3.12.1
+
+## Changes
+
+## Bugfixes
+
+- *Fixed tertiary stats not being shown in loot/voting frame.*
+- *Miscellaneous items is now listed as their subtype instead of "armor token" (unless their subtype is junk) in voting frame more info.*
+- *Setting "Require Note" on a button that has default text now actually makes said note required.*
+- *Fixed players guild ranks occasionally dissappearing.*
+
 # 3.12.0
 
 ## Changes